refactor(mobile): move named_resource_picker out of wizards/ layer
PR-5e precondition. The widget is now consumed by the ESO bulk-refresh
scope picker (mobile/lib/features/eso/...) as well as wizards, so it
no longer belongs under wizards/widgets/. Moves both the source and
the test file with git mv to preserve history.

Updates:
- 7 wizard import paths (../../widgets/ -> ../../../widgets/).
- Test import path + the relative support path.
- Doc comment on the source mentions the new ESO consumer.

flutter analyze clean.
