+++ mobile/lib/wizards/types/snapshot/snapshot_wizard_screen.dart
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../cluster/cluster_provider.dart';
-import '../../widgets/named_resource_picker.dart';
+import '../../../widgets/named_resource_picker.dart';
 import '../../widgets/wizard_review_body.dart';
 import '../../widgets/wizard_screen_scaffold.dart';
 import '../../widgets/wizard_unrouted_banner.dart';
